backend/requirements.txt
fastapi
pydantic
google-genai
python-dotenv
uvicorn
requests